2021[Rust all] Ran Clippy.
Should have done this earlier, fixed a few things but also sped up the
program overall. I need to integrate that into my regular flow.

+++ 2021/rust/src/day01.rs
@@ -9,7 +9,7 @@ pub fn part1() -> i64 {
     reader
         .lines()
         .map(|l| l.unwrap().parse::<i64>().unwrap())
-        .fold((0 as i64, i64::MAX), |(sum, prev), curr| {
+        .fold((0_i64, i64::MAX), |(sum, prev), curr| {
             (sum + if prev < curr { 1 } else { 0 }, curr)
         })
         .0
@@ -23,7 +23,7 @@ pub fn part2() -> i64 {
         .lines()
         .map(|l| l.unwrap().parse::<i64>().unwrap())
         .fold(
-            (0 as i64, (i64::MAX, i64::MAX, i64::MAX)),
+            (0_i64, (i64::MAX, i64::MAX, i64::MAX)),
             |(sum, (prev, b, c)), curr| (sum + if prev < curr { 1 } else { 0 }, (b, c, curr)),
         )
         .0
